+++ app/controllers/spina/primer_theme/conferences/application_controller.rb
@@ -1,8 +1,8 @@
 # frozen_string_literal: true
 
 module Spina
-  module Conferences
-    module PrimerTheme
+  module PrimerTheme
+    module Conferences
       # Base class from which controllers inherit
       class ApplicationController < Spina::ApplicationController
         def cookies_info
