+++ app/views/layouts/spina/primer_theme/journal/application.html.haml
