+++ app/views/layouts/spina/primer_theme/journal/articles.html.haml
