+++ app/views/layouts/spina/primer_theme/journal/issues.html.haml
