+++ app/views/primer_theme/pages/information.html.haml
