+++ app/views/spina/primer_theme/conferences/conferences/_conference.html.haml
