+++ app/views/spina/primer_theme/conferences/conferences/_event.html.haml
