+++ app/views/spina/primer_theme/conferences/conferences/_institution_logo.html.haml
