+++ app/views/spina/primer_theme/conferences/conferences/_presentation_type.html.haml
