+++ app/views/spina/primer_theme/conferences/conferences/index.html.haml
