+++ app/views/spina/primer_theme/conferences/kaminari/_next_page.html.haml
