+++ app/views/spina/primer_theme/conferences/kaminari/_page.html.haml
