+++ app/views/spina/primer_theme/conferences/kaminari/_paginator.html.haml
