+++ app/views/spina/primer_theme/conferences/kaminari/_prev_page.html.haml
