+++ app/views/spina/primer_theme/conferences/presentations/_abstract.html.haml
