+++ app/views/spina/primer_theme/journal/articles/show.html.haml
