+++ app/views/spina/primer_theme/journal/issues/_article.html.haml
