+++ app/views/spina/primer_theme/journal/issues/_issue.html.haml
