+++ app/views/spina/primer_theme/journal/issues/_issue_cover.html.haml
