+++ app/views/spina/primer_theme/journal/issues/index.html.haml
