+++ app/views/spina/primer_theme/journal/issues/show.html.haml
