--- lib/spina/conferences/primer_theme/version.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-module Spina
-  module Conferences
-    module PrimerTheme
-      VERSION = '0.4.2'
-    end
-  end
-end
