+++ lib/spina/primer_theme.rb
@@ -1,8 +1,8 @@
 # frozen_string_literal: true
 
 require 'spina'
-require 'spina/conferences/primer_theme/engine'
-require 'spina/conferences/primer_theme/breadcrumbs/builder'
+require 'spina/primer_theme/engine'
+require 'spina/primer_theme/breadcrumbs/builder'
 require 'spina/admin/conferences'
 require 'spina/admin/journal'
 require 'octicons_helper'
@@ -11,9 +11,7 @@ require 'primer/view_components'
 require 'icalendar'
 
 module Spina
-  module Conferences
-    module PrimerTheme
-      # Your code goes here...
-    end
+  module PrimerTheme
+    # Your code goes here...
   end
 end
