+++ lib/spina/primer_theme/version.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+module Spina
+  module PrimerTheme
+    VERSION = '0.4.2'
+  end
+end
